% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/codeCollection.R
\docType{data}
\name{ATCKoodit}
\alias{ATCKoodit}
\title{ATC codes}
\format{
A data frame:
\describe{
\item{ Koodi}{ Code value }
\item{ en }{ Label inEnglish}
\item{ fi }{ Label in English}
}
}
\usage{
ATCKoodit
}
\description{
A data set containing ATC codes
”@note Only in English
}
\keyword{datasets}
