% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/codeCollection.R
\docType{data}
\name{ATCKooditV2022}
\alias{ATCKooditV2022}
\title{ATC codes, V2022}
\format{
A data frame:
\describe{
\item{ Koodi}{ Code value }
\item{ en }{ Label inEnglish}
\item{ fi }{ Label in English}
}
}
\usage{
ATCKooditV2022
}
\description{
A data set containing ATC codes, updated 2022
”@note Only in English
}
\keyword{datasets}
