% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/codeCollection.R
\name{comb.levels}
\alias{comb.levels}
\title{Groups rare levels of factor}
\usage{
comb.levels(x, N.max = 10, lang = "fi")
}
\arguments{
\item{x}{factor or other categorical variable}

\item{N.max}{maximum number of levels}

\item{lang}{language ("fi" or "en")}
}
\value{
factor containing code value and text
}
\description{
Groups rare levels of factor
}
