% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/codeCollection.R
\docType{data}
\name{ICD10Koodit}
\alias{ICD10Koodit}
\title{Codebook for ICD10 codes}
\format{
A data frame:
\describe{
\item{ Koodi}{ Code value }
\item{ fi }{ Label in Finnish }
\item{ la }{ Label in Latin }
\item{ en }{ Name in English}
\item{ sv }{ Name in Swedish}
}
}
\usage{
ICD10Koodit
}
\description{
A data set containing ICD10 codes (without dots) and labels
(\url{https://koodistopalvelu.kanta.fi/codeserver/pages/classification-list-page.xhtml}).
}
\examples{
# ICD10 codes starting with I67
ICD10Koodit[grepl("^I67",ICD10Koodit$Koodi),]
}
\keyword{datasets}
