% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/codeCollection.R
\docType{data}
\name{ICDO3Koodit}
\alias{ICDO3Koodit}
\title{ICD-O-3 codes}
\format{
A data frame:
\describe{
\item{ Koodi}{ Code value }
\item{ fi }{ Label in English }
\item{ en }{ Label in English }
\item{ se }{ Label in English }
}
}
\usage{
ICDO3Koodit
}
\description{
A data set containing ICD-O-3 codes and labels
}
\keyword{datasets}
