% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/codeCollection.R
\docType{data}
\name{ICPCKoodit}
\alias{ICPCKoodit}
\title{ICPC codes}
\format{
A data frame:
\describe{
\item{ Koodi}{ Code value }
\item{ en }{ Label inEnglish}
\item{ fi }{ Label in English}
}
}
\usage{
ICPCKoodit
}
\description{
A data set containing International Classification of Primary Care (ICPC) codes.
If English code is missing uses Finnish, and vice versa (\url{https://koodistopalvelu.kanta.fi/codeserver/pages/classification-list-page.xhtml}).
}
\keyword{datasets}
