% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/codeCollection.R
\name{mkcodelabel}
\alias{mkcodelabel}
\title{Add label text to code}
\usage{
mkcodelabel(x, codebook, lang = "fi", as.factor.B = TRUE)
}
\arguments{
\item{x}{code value (character)}

\item{codebook}{data frame containing codes and text}

\item{lang}{language used, possible values "fi", "en", and "sv"}

\item{as.factor.B}{should value be converted to factor}
}
\value{
character vector of factor containing code value and text
}
\description{
Add label text to code
}
\examples{
mkcodelabel("CJB10",TPKoodit,"en",FALSE)
mkcodelabel("CJB10",TPKoodit)
mkcodelabel("I351",ICD10Koodit,"en",FALSE)
}
\author{
Jari Haukka \email{jari.haukka@helsinki.fi}
}
