% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/codeCollection.R
\docType{data}
\name{SPATKoodit}
\alias{SPATKoodit}
\title{SPAT procedure codes for primary care}
\format{
A data frame:
\describe{
\item{ Koodi}{ Code value }
\item{ en }{ Label in Finnish}
\item{ fi }{ Label in Finnish}
\item{ se }{ Label in Swedish}
}
}
\usage{
SPATKoodit
}
\description{
SPAT procedure codes for primary care (\url{https://koodistopalvelu.kanta.fi/codeserver/pages/classification-list-page.xhtml}).
”@note Only in Finnish and Swedish
}
\keyword{datasets}
