% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/codeCollection.R
\docType{data}
\name{TPKoodit}
\alias{TPKoodit}
\title{Codebook for NOMESCO procedure codes}
\format{
A data frame:
\describe{
\item{ Koodi}{ Code value }
\item{ fi }{ Label in Finnish }
\item{ en }{ Name in English}
\item{ sv }{ Name in Swedish}
}
}
\usage{
TPKoodit
}
\description{
A data set containing NOMESCO (\url{http://nowbase.org/publications/ncsp-classification-surgical-procedures}) procedure codes and labels
(\url{https://koodistopalvelu.kanta.fi/codeserver/pages/classification-list-page.xhtml}).
}
\examples{
# Procedure codes starting with CHD
TPKoodit[grepl("^CHD2",TPKoodit$Koodi),]
}
\keyword{datasets}
