% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/codeCollection.R
\docType{data}
\name{tpksaKoodit}
\alias{tpksaKoodit}
\title{Statistics Finland causes of death time series codes}
\format{
A data frame:
\describe{
\item{ Koodi}{ Code value }
\item{ en }{ Label in English}
\item{ fi }{ Label in Finnish}
}
}
\usage{
tpksaKoodit
}
\description{
Statistics Finland causes of death time series codes
( \url{https://stat.fi/en/luokitukset/kuolinsyyt/kuolinsyyt_1_20210101/} )
”@note Only in English
}
\keyword{datasets}
